+++ picoquic/tls_api.c
@@ -1775,7 +1775,7 @@ int picoquic_set_tls_root_certificates(picoquic_quic_t* quic, ptls_iovec_t* cert
     ptls_context_t* ctx = (ptls_context_t*)quic->tls_master_ctx;
     ptls_openssl_verify_certificate_t* verify_ctx = (ptls_openssl_verify_certificate_t*)ctx->verify_certificate;
 
-    for (int i = 0; i < count; ++i) {
+    for (size_t i = 0; i < count; ++i) {
         X509* cert = d2i_X509(NULL, (const uint8_t**)&certs[i].base, certs[i].len);
 
         if (cert == NULL) {
